producer/requirements.txt
kafka-python

requests

